fix: report real graph freshness timestamps
