+++ packages/opcore-graph-core-darwin-arm64/metadata.json
@@ -4,6 +4,6 @@
   "targetPlatform": "darwin-arm64",
   "binaryPath": "opcore-graph-core",
   "checksumPath": "opcore-graph-core.sha256",
-  "checksumSha256": "f08fa2afab6c6ff6221cfd5ad4fc5cd25ad744fa64773d46fe53f926c4493272",
+  "checksumSha256": "6e65d63da212130d1a3ad969ec37d609d31ed1a873404734814b52c407c4300c",
   "buildProfile": "release"
 }
